requirements.txt
pdfplumber
openai
pandas
openpyxl
python-dotenv
pytest
gradio
matplotlib
faiss-cpu
